Make View instance respond to model_name
In order to have duck-typing-like compatibility with
User::ActiveRecord_Relation this adds a
CouchRest::Model::Designs::View#model_name method that maps to the
parent model in View#model.

For my specific use-case I'm adding this for better out-of-the-box
compatibility with Pundit, but I'm sure that there are other gems that
could benefit from this as well.